ForeignerJobRecommendations - optional municipalityCode (#71)
This adds a v1.2 of the Employment/ForeignerJobRecommendations that has the municipalityCode in the jobs in the response set as optional, so it will allow null values.

Also fixed a typo in the description field (trailing space). This is fixed in all the previous versions of the definition as well.

This also marks the v1.1 as deprecated.