Fix typo in definition name and set version to appropriate level
This copies the `Employment/ForeignerJobRecommendatations_v1.3` [sic] to `Employment/ForeignerJobRecommendations_v0.2`. This thus fixes the typo in the word recommendations. The old v1.3 is marked as deprecated. At same time the new, correctly named definition, gets the version number reset to pre 1.0 to show it's still in pretty early stage, as can be seen from the bunch of updates made to it this week. Once it's been tested out more we can increase the version to 1.0.

The version number (0.2) might be a bit arbitrary, but this is on the other hand already iterated a few times and this would allow us if needed to also publish it on the DataFinland dataspace and there easily distinguish it from the one with the typo which is there called 0.1.